% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableize.R
\name{renderTableize}
\alias{renderTableize}
\title{Widget render function for use in Shiny}
\usage{
renderTableize(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

