% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableize.R
\name{tableize}
\alias{tableize}
\title{<Add Title>}
\usage{
tableize(d, fixedCols = NULL, rowLabelCol = NULL, fixedRows = NULL,
  header = NULL, style = "table0", placeholderRow = NULL,
  placeholderCol = NULL, selectedRows = NULL, selectedCols = NULL,
  width = NULL, height = NULL)
}
\description{
<Add Description>
}

